+++ lib/client.ts
@@ -71,9 +71,9 @@ export class WebSocket {
   public on(event: string, listener: (...args: any[]) => void): void {
     if (typeof listener !== 'function') {
       throw new Error(`Could not set listener for '${event}' event, listener must be a function`);
-    }
-
-    if (this.registeredEvents[event] !== noop) {
+    } else if (this.registeredEvents[event] === undefined) {
+      console.warn(`WebSocket Client does not support '${event}' listener`);
+    } else if (this.registeredEvents[event] !== noop) {
       throw new Error(`Can not set '${event}' event listener twice`);
     }