+++ tests/server-client.test.ts
@@ -301,4 +301,13 @@ describe('Server & Client', () => {
 
     const socket: WebSocket = new WebSocket('ws://localhost:3000');
   });
+
+
+  it('Should trigger on `close` event at the server', (done: any) => {
+    this.wsServer.on('connection', (connection: WebSocket) => { });
+    this.wsServer.on('close', () => {
+      done();
+    });
+    this.wsServer.close();
+  });
 });